Extract TypeLink out of DocExplorer
